requirements.txt
asgiref==3.7.2
Django==4.2.4
django-appconf==1.0.5
django-compressor==4.4
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.5.18
Pillow==10.0.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
tzdata==2023.3
